+++ CHANGES.rst
@@ -31,6 +31,10 @@ Future / To-Do
 Current
 -------
 -   `math/`
+    -   `math_core.py`
+        -   `minmax()`
+            -  Improved how 'stretch' is handled, and added separate 'log_stretch' parameter to stretch in log-space (as apposed to linear).
+        
     -   `statistic.py`
         -   `percentiles`
             -   BUG: issue with data type incompatibilities between input data and the percentiles.